+++ src/SDKs/DataFactory/Management.DataFactory/changelog.md
@@ -1,5 +1,24 @@
 # Changelog for the Azure Data Factory V2 .NET SDK
 
+## Version 0.2.2-preview
+
+### Feature Additions
+  * Add SAP Cloud For Customer Source
+  * Add SAP Cloud For Customer Dataset
+
+## Version 0.2.1-preview
+
+### Feature Additions
+  * Cancel pipeline run api.
+  * Add AzureMySql linked service.
+  * Add AzureMySql table dataset.
+  * Add AzureMySql Source.
+  * Add Dynamics Sink
+  * Add SalesforceObject Dataset
+  * Add Salesforce Sink
+  * Add jsonNodeReference and jsonPathDefinition to JSONFormat
+  * Support providing Salesforce passwords and security tokens as SecureStrings or as secrets in Azure Key Vault.
+
 ## Version 0.2.0-preview
 
 ### Feature Additions